\name{ecoliLeucine}
\docType{instance}
\alias{ecoliLeucine}
\title{AffyBatch instance ecoliLeucine}
\description{
The {\it Leucine-responsive regulatory protein} (Lrp) is thought to regulate
the expression of multiple genes and operons in bacteria. Two strains
{\it lrp+} and {\it lrp-} were studied to investigate which genes are
subject to changes in expression levels.

The raw data were taken from CyberT website. The data for 8 chips (4
{\it lrp+} and 4 {\it lrp-}).
}
%\usage{data(affybatch.example)}
\format{An \code{\link[affy:AffyBatch-class]{AffyBatch}} of 8 samples.}
\references{
  She-pin Hung, Pierre Baldi, and G. Wesley Hatfield, J. Biol. Chem.,
  Vol. 277, Issue 43, 40309-40323, October 25, 2002

  \url{http://visitor.ics.uci.edu/genex/cybert/tutorial/Affy\%20Data/}
}
\seealso{\code{\link[affy]{AffyBatch-class}}}
}
\keyword{data}


